only require `sdk` property for vite plugin
