+++ source-code/sdk-js/src/config/config.ts
@@ -1,13 +1,11 @@
 import { type InlangConfig, type InlangConfigModule, setupConfig } from "@inlang/core/config"
 import { initialize$import } from '@inlang/core/environment'
 import fs from "node:fs/promises"
-import { SdkConfigInput, validateSdkConfig } from '../plugin/schema.js'
+import type { SdkConfigInput } from '../plugin/schema.js'
 
 export type InlangConfigWithSdkProps = InlangConfig & {
 	sdk?: SdkConfigInput
 }
-
-class InlangSdkConfigError extends Error { }
 
 export class InlangError extends Error { }
 
@@ -35,27 +33,10 @@ Node.js failed to resolve the URL. This can happen sometimes during development.
 	return setupConfig({ module, env })
 }
 
-function assertConfigWithSdk(config: InlangConfig | undefined): asserts config is InlangConfigWithSdkProps {
-	if (!config) {
-		throw new InlangSdkConfigError('Could not find `inlang.config.js` in the root of your project.`')
-	}
-
-	if (!('sdk' in config)) {
-		// TODO: link to docs
-		throw new InlangSdkConfigError('The `sdk` property is missing in your `inlang.config.js` file.`. Make sure to use the `sdkPlugin` in your `inlang.config.js` file.')
-	}
-}
-
 export const initConfig = async (module: InlangConfigModule) => {
 	if (!module) {
 		throw Error("could not read `inlang.config.js`")
 	}
 
-	const config = await setupInlangConfig(module)
-
-	assertConfigWithSdk(config)
-
-	validateSdkConfig(config.sdk)
-
-	return config
+	return setupInlangConfig(module)
 }
